LayoutTests/http/wpt/webauthn/public-key-is-user-verifying-platform-authenticator-available.html

<!DOCTYPE html>
<title>Web Authentication API: PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable().</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    promise_test(function(t) {
        internals.mockAuthenticatorCoordinator;

        return PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable().then(function(result) {
            assert_equals(result, false);
        });
    }, "PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable() without any available authenticators.");

    promise_test(function(t) {
        internals.mockAuthenticatorCoordinator.setDidUserVerifyingPlatformAuthenticatorPresent();

        return PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable().then(function(result) {
            assert_equals(result, true);
        });
    }, "PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable() with available authenticators.");
</script>
